backend/requirements.txt
fastapi==0.115.7
firebase_admin==6.6.0
passlib==1.7.4
pydantic==2.10.5
PyJWT==2.10.1
pytest==8.3.4
python-dotenv==1.0.1
Requests==2.32.3
